anadrofish: Anadromous Fish Population Responses to Habitat Changes

The anadrofish package is a collection of tools for running coastwide or river-specific population models for anadromous fish in response to habitat change from dams.

Functions called directly

The following functions can be called directly from anadrofish:

sim_pop main function used to simulate populations add_unequal_vectors add vectors of unequal length beverton_holt Beverton-Holt recruitment with density dependence custom_habitat_template make custom habitat for an existing river or a template for a new river of interest get_dams get dams for specified river from American shad habitat data get_govt get governmental unit for specified river by species get_region get region for specified river by species get_rivers find out which rivers are available make_downstream river-specific, catchment-wide downstream survival through dams make_eggs simulate eggs per female by species and river make_habitat subset of habitat dataset for selected river make_iteroparity predict iteroparity from latitude based on reported relationships make_lat get latitude for specified river make_maxage get region-specific maximum age for specified population make_mortality calculate natural, instantaneous mortality by river make_pop simulate starting population make_postspawn predict post-spawning survival based on natural mortality and iteroparity make_recruits predict recruits from number of adults, fecundity, sex ratio, and juvenile survival make_spawners draw spawners from age-specific spawn recruitment probabilities make_spawnrecruit get region- and age-specific probabilities of recruitment to spawn project_pop project population to next time step without reproduction sim_juvenile_s simulate juvenile survival based on reported rates lower95 convenience function for calculating upper 95% CI upper95 convenience function for calculating lower 95% CI

Data

The following built-in datasets are included:

crecco_1983 Regression parameters for latitude-iteroparity relationship in American shad fl_tl_conversions Fork length - total length conversions for river herring habitat American shad habitat data habitat_ale Alewife habitat data habitat_bbh Blueback herring habitat data jessop_1993 Fork length-fecundity relationships for alewife length_weight Regional length-weight regression parameters for American shad lw_pars_rh Regional length-weight regression parameters for river herring maki_pars Regional spawner recruitment parameters for river herring maturity Regional spawner recruitment probabilities for American shad max_ages Regional maximum ages by sex and species mortality Regional mortality estimates for American shad by sex mortality_rh Regional mortality estimates for river herring by sex olney_mcbride Regression parameters for weight-batch fecundity relationships vbgf_ale von Bertalanffy growth parameters for alewife vbgf_bbh von Bertalanffy growth parameters for blueback herring vbgf_NI von Bertalanffy growth parameters for NI American shad vbgf_SI von Bertalanffy growth parameters for SI American shad vbgf_SP von Bertalanffy growth parameters for SP American shad

See also

Author

Maintainer: Daniel S. Stich daniel.stich@oneonta.edu (ORCID)

Authors:

S. G. Roy

S. D. Snyder

J. D. Zydlewski