Daily mortality of juvenile American shad

A built-in dataset containing estimated daily mortality rates for larval and juvenile American shad in the Connecticut River, USA 1979-1982 (Crecco et al. 1983).

Usage

crecco_1983

Format

A data frame with 24 observations of 6 variables:

Year

Year of study

Age

Fish age (d)

Md

Daily mortality

Sd

Daily survival

Si

Interval survival

Sc

Cumulative survival

References

Crecco, V., T. Savoy, and L. Gunn. 1983. Daily mortality rates of larval and juvenile American shad (*Alosa sapidissima*) in the Connecticut River with changes in year-class strength. Canadian Journal of Fisheries and Aquatic Sciences 40:1719–1728.