Fill output vectors

Internal function to fill output vectors for default arguments in sim_pop and the output (spawners).

For internal use in sim_pop. Not intended to be called directly.

Usage

fill_output(.sim_pop, sex_specific = FALSE)

Arguments

.sim_pop

A hidden environment in the calling frame of sim_pop. Arbitrarily any named object with names matching those used in the function.

sex_specific

Logical inherited from sim_pop indicating whether to use sex-specific output.

Value

list of output vectors