Alewife habitat

A dataset containing the surface area of Alewife habitat and features for individual flow line segments or waterbodies (lakes) comprising habitat units in Atlantic coast drainages.

Usage

habitat_ale

Format

A data frame with 246,259 observations of 7 variables:

REACHCODE code used to identify individual habitat units Hab_sqkm surface area of habitat in square kilometers Latitude latitude at downstream terminus of habitat unit State governmental unit at downstream terminus of habitat unit River_huc name of river to which habitat unit belongs, derived from hydrologic unit codes POP genetic reporting group geographic region within which habitat unit falls DamOrder order of dam at downstream terminus of habitat unit. Cumulatively assigned such that all habitat units upstream of a given dam all have dam_order >= 1

Source

Shawn Snyder

Examples

head(habitat_ale)
#>   REACHCODE    Hab_sqkm Latitude State                           River_huc POP
#> 1  2.04E+12 0.008199607 41.12328    NJ                      Upper Delaware MAT
#> 2  2.04E+12 0.010045756 39.88029    PA                      Lower Delaware MAT
#> 3  1.09E+12 0.016462307 42.12254    MA Massachusetts-Rhode Island Coastal. SNE
#> 4  1.10E+12 0.141209610 41.38672    CT                 Connecticut Coastal SNE
#> 5  2.02E+12 0.006305018 41.30567    NY                        Upper Hudson MAT
#> 6  1.09E+12 0.000186516 42.27801    MA Massachusetts-Rhode Island Coastal. SNE
#>   DamOrder
#> 1        7
#> 2        0
#> 3        5
#> 4        1
#> 5        3
#> 6        8