Length-weight regression coefficients for American shad

A dataset containing sex-specific regression coefficients of log10length-log10weight relationships for each life-history region (NI = Northern iteroparous, SI = Southern iteroparous, SP = Semelparous). Use to predict weight at age from length derived using region-specific von Bertalanffy growth parameters in built-in datasets.

Usage

length_weight

Format

A data frame with 6 observations of 4 variables:

region Life-history region sex Fish sex F is female and M is male alpha Intercept beta Slope

Source

Atlantic States Marine Fisheries Commission

References

Atlantic States Marine Fisheries Commission (ASMFC). 2020. American shad benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.