Simulate number of eggs per fish for river herring

Function used to simulate number of eggs produced per female (total fecundity) for river herring in rivers included in get_rivers.

Usage

make_eggs_rh(river, species = c("ALE", "BBH"), custom_habitat = NULL)

Arguments

river

The river for which eggs will be simulated.

species

Species for which population dynamics will be simulated. Choices include alewife ("ALE"), and blueback herring ("BBH").

custom_habitat

A dataframe containing columns corresponding to the those in the output from custom_habitat_template(). NEED TO ADD LINK.

Value

A vector containing age-specific potential annual fecundity with length = max_age.

Details

NEED TO ADD DETAILS.

References

Sullivan et al. (2019), Jessop (1993)

Examples

make_eggs_rh(river = "Upper Susquehanna", species = "BBH")
#>  [1]  65215.36 149340.66 217867.63 263289.02 290644.45 306344.98 315132.67
#>  [8] 319985.91 322647.07 324100.61 324892.87 325324.21