Sex-specific von Bertalanffy growth parameters for alewife

A dataset containing sex-specific estimates of von Bertalanffy growth parameters for alewife from the 2024 ASMFC river herring benchmark stock assessment (ASMFC 2024).

Usage

vbgf_ale

Format

A data frame with 1,116,000 observations of 5 variables:

Region Genetic reporting group Sex Sex of fish Linf Asymptotic length of fish K Brody growth coefficient t0 Size at age zero

Source

Atlantic States Marine Fisheries Commission

References

Atlantic States Marine Fisheries Commission. 2024. River herring benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf