Make total spawning migration mortality

Function used to derive total in-river adult survival during the spawning migration from annual mortality and post-spawn survival.

Usage

make_s_spawn(nM, s_postspawn)

Arguments

nM

Instantaneous natural mortality rate. A numeric vector containing either an age-invariant natural mortality rate of length 1 or an age-specific natural mortality rate with length = max_age.

s_postspawn

Post spawning mortality calculated from the output of make_iteroparity and nM, or alternatively a numeric vector for each.

Value

A vector of length one containing pre-spawn survival.

Examples

make_s_spawn(0.68, 0.92)
#> [1] 0.8214248